Adding decorator based deploy package team_user and demo_app
