+++ demo_app/defaults.py
@@ -1,5 +1,6 @@
 DEFAULTS={
         'demo_app_user' : 'devops',
         'demo_app_user_home' : '/home/devops',
+        'settings': True,
         }
