+++ demo_app/demo_app.py
@@ -4,15 +4,16 @@ from pyinfra import local
 
 from pyinfra import logger
 
-from utils import get_file_path
-from utils import get_template_path
-from utils import get_task_path
-
 from .defaults import DEFAULTS
 
 
 @deploy('Setup user account', data_defaults=DEFAULTS)
 def setup_app(
-    state,host
-    ):
-    pass
+   state,host,ex_settings=None
+   ):
+   logger.info("ex_settings initial value : {}".format(ex_settings))
+   if (not ex_settings) and host.data.settings:
+      ex_settings = host.data.settings
+
+   logger.info("ex_settings after value : {}".format(ex_settings))
+
