+++ deploy_demo.py
@@ -1,9 +1,10 @@
-from team_user.team_user import setup_user
-from demo_app.demo_app import setup_app
+
 from pyinfra import logger
 from pyinfra import host
 
-logger.info("Host is {}:".format(host.fact.os))
+from team_user.team_user import setup_user
+from demo_app.demo_app import setup_app
 
+logger.info(host.fact.os)
 setup_user()
-setup_app()
+setup_app(ex_settings=False)
