+++ team_user/utils.py
@@ -14,13 +14,20 @@ from botocore.exceptions import NoCredentialsError
 def get_template_path(file):
     return path.join(
         path.dirname(__file__),
-        'templates',
-        file,
-    )
-def get_file_path(file):
-    return path.join(
-        path.dirname(__file__),
-        'files',
+        "templates",
         file,
     )
 
+def get_file_path(file):
+    return path.join(
+        path.dirname(__file__),
+        "files",
+        file,
+    )
+
+def get_task_path(file):
+    return path.join(
+        path.dirname(__file__),
+        "tasks",
+        file,
+    )
